pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "toonprompt"
version = "0.1.0"
description = "Universal prompt optimization wrapper for coding CLIs that compresses structured context into Toon-style blocks."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name = "OpenAI Codex"}
]
dependencies = [
  "PyYAML>=6.0",
  "tomli>=2.0; python_version < '3.11'"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0"
]

[project.scripts]
toon = "toonprompt.cli:main"
toon-codex = "toonprompt.cli:main_codex"
toon-claude = "toonprompt.cli:main_claude"
toon-cursor = "toonprompt.cli:main_cursor"
toon-gemini = "toonprompt.cli:main_gemini"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
